refactor(reindex): the loop is a module, not a route
Last of the three extractions B-2 needed, and the only one where nothing existed to
wrap: the re-embedding work was written inline in the route handler as five
near-identical batch loops, each with its own projection, its own *EmbedText builder
and its own per-record error tolerance. 784 -> 588 lines in the router.

`brain/reindex.ts`: planReindex decides (404, the proxy 400, the single-job 409) and
startReindex runs. `memberIds` is passed IN rather than resolved there, because scope
resolution differs per surface -- REST narrows by request, MCP by the token's
accessible spaces -- and re-deriving it inside would be a second place for the two to
disagree about which spaces a token may touch.

The guard and the metric moved WITH the work, and that is the part that matters.
`reindexJobRunning` was module state in the router. A second surface calling the loop
directly would have run a concurrent job the guard could not see, and
`reindexInProgress` would have been left at 1 by whichever job finished second -- so
the flag would report a reindex running forever, on an instance with none.

startReindex returns as soon as the job is SCHEDULED and never awaits it, keeping the
contract the route already had: `status: started` with zeroed counters, work on the
next turn, headers flushed immediately. Awaiting it would answer the same 200 and turn
a multi-minute job into a request timeout.

Two gates got STRONGER rather than merely re-pointed, which is the outcome to aim for
when a check follows moved code:

  - the proxy gate asserted `res.status(400)` appeared before `reindexJobRunning =
    true`. It now asserts the decision function never sets the flag at ALL, so no
    refusal can leave it set, plus that startReindex is where the guard is taken. A
    position comparison replaced by an absence, which cannot be satisfied by reordering.
  - the builder gate reads the module now, so on its own it would pass for a route that
    re-grew its own loop. It gained the other half: the route must call startReindex and
    must build no embed text itself.

Also re-pointed: the document-description excerpt check and the reindex-refuses-a-proxy
slice. The dead `*EmbedText` import left behind in the router was removed rather than
left to rot -- it is exactly the kind of leftover that makes the next reader think the
route still embeds.

Verified serially on a rebuilt image: 237 tests, 0 failures -- the reindex contract
suite that landed against the unmoved route (#853), plus brain and proxy-spaces.

The tool follows in its own PR. When it lands, REST_ONLY_CAPABILITIES is empty.
